Add BookList module
