+++ index.html
@@ -68,7 +68,7 @@
   <footer>
     <p tabindex="0"> Julio Quezada - Full Stack Web Developer &copy; 2023</p>
   </footer>
-  <script src="./index.js"></script>
+  <script type="module" src="./index.js"></script>
 </body>
 
 </html>
